finish v1 of create memento endpoint
