+++ backend/pyproject.toml
@@ -36,6 +36,7 @@ keras = "^3.8.0"
 onnxruntime = "^1.20.1"
 numba = "0.61.0"
 rembg = {version = "2.0.61", extras = ["cpu"]}
+python-multipart = "^0.0.20"
 
 
 [tool.poetry.group.dev.dependencies]
